pyproject.toml
[tool.poetry]
name = "document_analisis"
version = "0.1.0"
description = "document_analisis using crewAI"
authors = ["Elton es18423772gmail.com"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
python-dotenv = "^1.0.0"
crewai = "^0.32.2"
crewai-tools = "^0.3.0"
translate = "^3.6.1"
httpx = "^0.27.0"
openai = ">=1.24.0,<2.0.0"
nltk = "^3.6.3"
spacy = "^3.1.0"
pandas = "^1.3.0"
flask = "^2.0.1"
PyPDF2 = "^1.26.0"
python-docx = "^0.8.11"
pymupdf = "^1.18.13"
asyncio = "^3.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
